# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private town-square API for a residential community.",
        "title": "Cul-de-Chat API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/directory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Active users who opted in. Shows name and unit number.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Resident directory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.DirectoryUserDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/media/profile/{filename}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Fetch an uploaded profile photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
        "/profile/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get my profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ProfileDTO"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update my profile",
                "parameters": [
                    {
                        "description": "Profile fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateProfileInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ProfileDTO"
                        }
                    }
                }
            }
        },
        "/profile/me/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Upload profile photo",
                "parameters": [
                    {
                        "type": "file",
                        "description": "jpeg, png, or webp, max 2MB",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ProfileDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "services.DirectoryUserDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                },
                "unit_number": {
                    "type": "string"
                }
            }
        },
...
        "services.ProfileDTO": {
            "type": "object",
            "properties": {
                "directory_opt_in": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                },
                "unit_number": {
                    "type": "string"
                }
            }
        },
...
        "services.UpdateProfileInput": {
            "type": "object",
            "properties": {
                "directory_opt_in": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}